+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+- New `Session` module to fetch and set session values
+- New `cookie_session` middleware that persists session data in cookie
 - New `Auth` module to work with `Authorization` header (#238)
 - New `basic_auth` middleware to protect handlers with a `Basic` authentication method (#238)